fix: change file namings, fix typescript support for styled-components
