+++ src/components/App/App.tsx
@@ -1,0 +1,5 @@
+import { StyledParagraph } from "./styled.ts";
+
+export const App = () => {
+    return <StyledParagraph>App Hello world</StyledParagraph>;
+};
