+++ src/components/LayoutContainer/LayoutContainer.tsx
@@ -1,0 +1,5 @@
+import { StyledContainer } from "./styled.ts";
+
+export const LayoutContainer = ({ children }: { children: React.ReactNode }) => {
+    return <StyledContainer>{children}</StyledContainer>;
+};
