+++ src/context/StyledComponentsProvider/index.tsx
@@ -1,12 +1,3 @@
-import { GlobalStyles } from "@styles/globalStyles";
-import { darkTheme } from "@styles/theme";
-import { ThemeProvider } from "styled-components";
+import { StyledComponentsProvider } from "./StyledComponentsProvider.tsx";
 
-export function StyledComponentsProvider({ children }: { children: React.ReactNode }) {
-    return (
-        <ThemeProvider theme={darkTheme}>
-            <GlobalStyles />
-            {children}
-        </ThemeProvider>
-    );
-}
+export default StyledComponentsProvider;
