+++ src/index.tsx
@@ -1,5 +1,5 @@
-import { App } from "@components/App";
-import { StyledComponentsProvider } from "@context/StyledComponentsProvider";
+import App from "@components/App";
+import StyledComponentsProvider from "@context/StyledComponentsProvider";
 import { createRoot } from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
