+++ src/pages/Bank/Bank.tsx
@@ -1,0 +1,7 @@
+export function Bank() {
+    return (
+        <>
+            <h1>Bank page</h1>
+        </>
+    );
+}
