+++ src/pages/Contacts/Contacts.tsx
@@ -1,0 +1,7 @@
+export function Contacts() {
+    return (
+        <>
+            <h1>Contacts page</h1>
+        </>
+    );
+}
