+++ src/pages/Home/Home.tsx
@@ -1,0 +1,7 @@
+export function Home() {
+    return (
+        <>
+            <h1>Home page</h1>
+        </>
+    );
+}
