+++ src/pages/Timeline/Timeline.tsx
@@ -1,0 +1,7 @@
+export function Timeline() {
+    return (
+        <>
+            <h1>Timeline page</h1>
+        </>
+    );
+}
