+++ tsconfig.json
@@ -27,5 +27,5 @@
             "@utils/*": ["./src/utils/*"]
         }
     },
-    "include": ["src/**/*"]
+    "include": ["src/**/*", "src/styles/styled.d.ts"]
 }
